+++ app/Config/Paths.php
@@ -75,4 +75,6 @@ class Paths
      * is used when no value is provided to `Services::renderer()`.
      */
     public string $viewDirectory = __DIR__ . '/../Views';
+
+    public string $supportDirectory = __DIR__ . '/../../vendor/codeigniter4/framework/system/ThirdParty';
 }